maj Ratings

+++ backend/controllers/books.js
@@ -1,30 +1,82 @@
 const Book = require('../models/Book');
+const Rating = require('../models/Rating');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
-    delete req.body._id;
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId
     const book = new Book({
-        ...req.body
+        ...req.body,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     });
     book.save()
         .then(() => res.status(201).json({ message: 'livre enregistré'}))
         .catch(error => res.status(400).json({ error }));
+// next;
+} 
+
+
+exports.createRating = (req, res, next) => {
+    const ratingsObject = JSON.parse(req.body.ratings);
+    delete ratingsObject._id;
+    delete ratingsObject._userId
+    const rating = new Rating({
+        ...req.ratings,
+        userId: req.auth.userId,
+        grade : req.grade.ratings
+
+    })
+    rating.save()
+    .then(() => res.status(201).json({ message: 'note enregistrée'}))
+    .catch(error => res.status(400).json({ error }));
 }
 
 exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id}, {...req.body, _id: req.params.id})
-        .then(() => res.status(200).json({ message: 'livre modifié'}))
-        .catch(error => res.status(400).json({ error }));
-}
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } : { ...req.body };
+  
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message : 'Non autorisé'});
+            } else {
+                Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+                .then(() => res.status(200).json({message : 'Livre modifié!'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
+        .catch((error) => {
+            res.status(400).json({ error });
+        });
+ };
 
-exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'livre supprimé'}))
-        .catch(error => res.status(400).json({ error }));
-}
+ exports.deleteBook = (req, res, next) => {
+    Book.findOne({ _id: req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({message: 'Non autorisé'});
+            } else {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({message: 'Livre supprimé !'})})
+                        .catch(error => res.status(401).json({ error }));
+                });
+            }
+        })
+        .catch( error => {
+            res.status(500).json({ error });
+        });
+ };
 
 exports.findOneBook = (req, res, next) => {
-    Book;findOne({ _id: req.params.id})
-        .then(thing => res.status(200).json(book))
+    Book.findOne({ _id: req.params.id})
+        .then(book => res.status(200).json(book))
         .catch(error => res.status(404).json({ error })); 
 }
 
@@ -33,3 +85,4 @@ exports.findAllBook = (req, res, next) => {
         .then(books => res.status(200).json(books))
         .catch(error => res.status(400).json({ error }));
 }
+
